+++ dabest/_api.py
@@ -2,35 +2,42 @@
 # -*-coding: utf-8 -*-
 # Author: Joses Ho
 # Email : joseshowh@gmail.com
+from __future__ import division
 
-
-def plot(data, idx,
+def plot(
+        data, idx,
         x=None, y=None,
         color_col=None,
-
-        float_contrast=True,
         paired=False,
-        show_pairs=True,
-        group_summaries="mean_sd",
+        id_col=None,
+        float_contrast=True,
+        fig_size=None,
 
-        custom_palette=None,
+        stat_func=None,
+        ci=95,
+        n_boot=5000,
+
+        swarm_dotsize=8,
+        difference_dotsize=10,
 
         swarm_label=None,
         contrast_label=None,
         swarm_ylim=None,
         contrast_ylim=None,
 
-        fig_size=None,
+        custom_palette=None,
 
-        font_scale=1.2,
-
-        stat_func=None,
-        ci=95,
-        n_boot=5000,
+        show_pairs=True,
         show_group_count=True,
-
-        tick_length=15,
-        tick_pad=9,
+        group_summaries="mean_sd",
+        context='talk',
+        ci_linewidth=3,
+        summary_linewidth=3,
+        multiplot_horizontal_spacing=0.75,
+        font_scale=1,
+        dpi=100,
+        tick_length=10,
+        tick_pad=5,
 
         swarmplot_kwargs=None,
         violinplot_kwargs=None,
@@ -44,6 +51,8 @@ def plot(data, idx,
     Takes a pandas DataFrame and produces a contrast plot:
     either a Cummings hub-and-spoke plot or a Gardner-Altman contrast plot.
     Paired and unpaired options available.
+
+
 
     Keywords:
         data: pandas DataFrame
@@ -59,16 +68,52 @@ def plot(data, idx,
         color_col: string, default None
             Column to be used for colors.
 
-        swarm_label, contrast_label: strings, default None
-            Set labels for the y-axis of the swarmplot and the contrast plot,
-            respectively.
+        paired: boolean, default False
+            Whether or not the data is paired. If True, `idx` must consist of a
+            tuple of length 2, or each tuple in `idx` must have a length of 2.
+            The first category/group is the observation before treatment, whilst
+            the second category/group is that after treatment.
+
+        id_col: string, default None
+            If `paired` is True, this must be supplied. This column indicates
+            the identity of the datapoint if the data is paired.
 
         float_contrast: boolean, default True
             Whether or not to display the halfviolin bootstrapped difference
             distribution alongside the raw data.
 
-        paired: boolean, default False
-            Whether or not the data is paired. To elaborate.
+        fig_size: tuple, default None
+            The desired dimensions of the figure as a (length, width) tuple.
+
+        stat_func: callable, default None
+            The function used to compute the summary. If None, defaults to
+            np.mean()
+
+        ci: integer, default 95
+            The size of the confidence interval desired (in percentage).
+            Sensible values range from 50 to 99.
+
+        n_boot: integer, default 5000
+            Number of bootstrap iterations to perform during calculation of
+            confidence intervals.
+
+        swarm_dotsize: integer, default 8
+            The size of the dots used to plot the rawdata in the swarmplot.
+
+        difference_dotsize: integer, default 10
+            The size of the dots used to indicate the effect sizes.
+
+        swarm_label, contrast_label: strings, default None
+            Set labels for the y-axis of the swarmplot and the contrast plot,
+            respectively.
+
+        swarm_ylim: tuple, default None
+            The desired y-limits of the raw data swarmplot as a (lower, higher)
+            tuple.
+
+        contrast_ylim: tuple, default None
+            The desired y-limits of the constrast plot as a (lower, higher)
+            tuple.
 
         custom_palette: dict, list, or matplotlib color palette, default None
             This keyword accepts a dictionary with {'group':'color'} pairings,
@@ -95,6 +140,10 @@ def plot(data, idx,
             swarmplot, or as paired plot, with a line joining each pair of
             observations.
 
+        show_group_count: boolean, default True
+            Whether or not the group count (e.g. 'N=10') will be appended to the
+            xtick labels.
+
         group_summaries: ['mean_sd', 'median_quartiles', 'None'], default 'mean_sd'
             Plots the summary statistics for each group. If 'mean_sd', then the
             mean and standard deviation of each group is plotted as a notched
@@ -102,47 +151,44 @@ def plot(data, idx,
             median and 25th and 75th percentiles of each group is plotted
             instead. If 'None', the summaries are not shown.
 
-        swarm_ylim: tuple, default None
-            The desired y-limits of the raw data swarmplot as a (lower, higher)
-            tuple.
+        context: string, default 'talk'
+            This implements the plotting context, as defined in seaborn.
+            There are four contexts: “paper”, “notebook”, “talk”, and “poster”,
+            which have plot elements (e.g. font scales) scaled by 0.8x, 1x,
+            1.3x, and 1.6x, respectively.
 
-        contrast_ylim: tuple, default None
-            The desired y-limits of the constrast plot as a (lower, higher)
-            tuple.
+        ci_linewidth: float, default 3
+            The width of the vertical line (in points) indicating the confidence
+            interval.
 
-        fig_size: tuple, default None
-            The desired dimensions of the figure as a (length, width) tuple.
+        summary_linewidth: float, default 3
+            The width of the vertical gapped line (in points) indicating the
+            summary of each group (as specified by `group_summaries`)
 
-        font_scale: float, default 1.2
-            The font size will be scaled by this number.
+        multiplot_horizontal_spacing: float, default 0.75
+            If the plot is a two-group multiplot with floating contrast axes,
+            this determines the horizontal spacing between each plot. This
+            number is a fraction of the axes width.
 
-        stat_func: callable, default None
-            The function used to compute the summary. If None, defaults to
-            np.mean()
+        font_scale: float, default 1.
+            The font size will be scaled by this number. This value is passed
+            on to seaborn.set().
 
-        ci: integer, default 95
-            The size of the confidence interval desired (in percentage).
+        dpi: float, default 100
+            The dots per inch of the figure produced. The default of 100 matches
+            the matplotlib default for both screen display and image saving.
 
-        n_boot: integer, default 5000
-            Number of bootstrap iterations to perform during calculation of
-            confidence intervals.
-
-        show_group_count: boolean, default True
-            Whether or not the group count (e.g. 'N=10') will be appended to the
-            xtick labels.
-
-        tick_length: int, default 15
+        tick_length: int, default 10
             The length of the ticks (in points) for both the swarm and contrast
             axes.
 
-        tick_pad: int, default 9
+        tick_pad: int, default 5
             The distance of the tick label from the tick (in points), for both
             the swarm and contrast axes.
 
         swarmplot_kwargs: dict, default None
             Pass any keyword arguments accepted by the seaborn `swarmplot`
-            command here, as a dict. If None, the following keywords are passed
-            to sns.swarmplot: {'size':10}.
+            command here, as a dict.
 
         violinplot_kwargs: dict, default None
             Pass any keyword arguments accepted by the matplotlib `
@@ -154,7 +200,8 @@ def plot(data, idx,
             This will change the appearance of the zero reference lines. Pass
             any keyword arguments accepted by the matplotlib Axes `hlines`
             command here, as a dict. If None, the following keywords are passed
-            to Axes.hlines: {'linestyle':'solid', 'linewidth':0.75, 'color':'k'}.
+            to Axes.hlines:
+            {'linestyle':'solid', 'linewidth':0.75, 'color':'k'}.
 
         group_summary_kwargs: dict, default None
             Pass any keyword arguments accepted by the matplotlib.lines.Line2D
@@ -167,12 +214,17 @@ def plot(data, idx,
             Pass any keyword arguments accepted by the matplotlib Axes `legend`
             command here, as a dict. If None, the following keywords are passed
             to Axes.legend:
-            {'loc': 'upper left', 'frameon': False, 'bbox_to_anchor': (0.95, 1.),
-            'markerscale': 2}.
+            {'loc': 'upper left', 'frameon': False, 'markerscale': 1.5,
+            'bbox_to_anchor': (0.95, 1.)}.
 
         aesthetic_kwargs: dict, default None
             Pass any keyword arguments accepted by the seaborn `set` command
-            here, as a dict.
+            here, as a dict. If None, the following keywords are passed to
+            seaborn.set:
+            {'context': 'poster', 'style': 'ticks', 'font_scale': font_scale}
+            (`font_scale` is a keyword as noted above with default 0.9.)
+
+
 
      Returns:
         A matplotlib Figure.
@@ -229,6 +281,7 @@ def plot(data, idx,
     import matplotlib.ticker as tk
     import matplotlib.lines as mlines
     from mpl_toolkits.axes_grid1 import make_axes_locatable
+    # Ensure that text will be saved as text in SVG files.
     plt.rcParams['svg.fonttype'] = 'none'
 
     import numpy as np
@@ -236,23 +289,24 @@ def plot(data, idx,
     import seaborn as sns
     import pandas as pd
 
-    from .plot_tools import halfviolin, align_yaxis, rotate_ticks
-    from .plot_tools import gapped_lines, get_swarm_spans
-    from .bootstrap_tools import bootstrap, jackknife_indexes, bca
-    from .misc_tools import merge_two_dicts, unpack_and_add
+    from ._plot_tools import halfviolin, align_yaxis, rotate_ticks
+    from ._plot_tools import gapped_lines, get_swarm_spans
+    from ._bootstrap_tools import bootstrap, jackknife_indexes, bca
+    from ._misc_tools import merge_two_dicts, unpack_and_add
 
     # MAKE COPY OF DATA.
     data_in = data.copy()
     data_in.reset_index(inplace=True)
 
-    # IDENTIFY PLOT TYPE.
+    # Determine the kind of estimation plot we need to produce.
     if all([isinstance(i, str) for i in idx]):
         plottype = 'hubspoke'
         # Set columns and width ratio.
         ncols = 1
+        ngroups = len(idx)
         widthratio = [1]
 
-        if len(idx) > 2:
+        if ngroups > 2:
             paired = False
             float_contrast = False
         # flatten out idx.
@@ -264,12 +318,14 @@ def plot(data, idx,
     elif all([isinstance(i, (tuple, list)) for i in idx]):
         plottype = 'multigroup'
         all_plot_groups = np.unique([tt for t in idx for tt in t]).tolist()
-        # Set columns and width ratio.
-        ncols = len(idx)
         widthratio = [len(ii) for ii in idx]
         if [True for i in widthratio if i > 2]:
             paired = False
             float_contrast = False
+        # Set columns and width ratio.
+        ncols = len(idx)
+        ngroups = len(all_plot_groups)
+
 
     else: # mix of string and tuple?
         err = 'There seems to be a problem with the idx you'
@@ -278,6 +334,18 @@ def plot(data, idx,
 
 
     # SANITY CHECKS
+    # If `paired` is True, check that `id_col` is supplied.
+    # Then check that `id_col` is actually a column in `data`.
+    # Lastly, sort `data` by `id_col`.
+    if paired is True:
+        if id_col is None:
+            raise ValueError("`paired` is True, but no `id_col` was given.")
+        if id_col not in data_in.columns:
+            err = "{} is not a column in `data`".format(id_col)
+            raise ValueError(err)
+        data_in.sort_values(id_col, inplace=True)
+
+
     # check color_col is a column name.
     if (color_col is not None) and (color_col not in data_in.columns):
         err = ' '.join(['The specified `color_col`',
@@ -309,6 +377,9 @@ def plot(data, idx,
         for g in all_plot_groups:
             if g not in data_in[x].unique():
                 raise IndexError('{0} is not a group in `{1}`.'.format(g, x))
+        # Filter the dataset so only the categories in flattened idx
+        # are considered.
+        data_in = data_in.loc[data_in[x].isin(all_plot_groups)]
 
     elif x is None and y is None:
         # Assume we have a wide dataset.
@@ -357,13 +428,14 @@ def plot(data, idx,
     # CALCULATE RAW SWARM YLIMS.
     if swarm_ylim is None:
         # To ensure points at the limits are clearly seen.
-        pad = data_in[y].diff().abs().min() / 2 #
-        swarm_ylim = (np.floor(data_in[y].min() - pad),
-                      np.ceil(data_in[y].max() + pad))
+        minimum, maximum = (data_in[y].min(), data_in[y].max())
+        # midpoint = data_in[y].mean()
+        pad = (maximum - minimum) * 0.1 # 10% of the data range.
+        swarm_ylim = (minimum - pad, maximum + pad)
 
 
     # Set default kwargs first, then merge with user-dictated ones.
-    default_swarmplot_kwargs = {'size':10}
+    default_swarmplot_kwargs = {'size': swarm_dotsize}
     if swarmplot_kwargs is None:
         swarmplot_kwargs = default_swarmplot_kwargs
     else:
@@ -395,8 +467,8 @@ def plot(data, idx,
 
     # Legend.
     default_legend_kwargs = {'loc': 'upper left', 'frameon': False,
-                             'bbox_to_anchor': (0.95, 1.),
-                             'markerscale': 2}
+                             'bbox_to_anchor': (1., 1.),
+                             'markerscale': 1.5}
     if legend_kwargs is None:
         legend_kwargs = default_legend_kwargs
     else:
@@ -404,9 +476,8 @@ def plot(data, idx,
 
 
     # Aesthetic kwargs for sns.set().
-    default_aesthetic_kwargs={'context': 'poster',
-        'style': 'ticks',
-        'font_scale': font_scale}
+    default_aesthetic_kwargs={'context': context, 'style': 'ticks',
+                              'font_scale': font_scale}
     if aesthetic_kwargs is None:
         aesthetic_kwargs = default_aesthetic_kwargs
     else:
@@ -423,7 +494,7 @@ def plot(data, idx,
         raise ValueError('group_summaries must be one of'
         'these: {}.'.format(gs_default) )
 
-    default_group_summary_kwargs = {'zorder': 5, 'lw': 4,
+    default_group_summary_kwargs = {'zorder': 5, 'lw': summary_linewidth,
                                     'color': 'k','alpha': 1}
     if group_summary_kwargs is None:
         group_summary_kwargs = default_group_summary_kwargs
@@ -448,13 +519,13 @@ def plot(data, idx,
     sns.set(**aesthetic_kwargs)
 
     swarm_xspan = 3
-    float_spacing = 1
+    #float_spacing = 1
 
     # Set appropriate horizontal spacing between subplots,
     # based on whether the contrast is floating.
     if float_contrast is True:
         # Check if this is multi two-group or not.
-        ws = float_spacing
+        ws = multiplot_horizontal_spacing
     else:
         ws = 0
 
@@ -465,37 +536,39 @@ def plot(data, idx,
     else:
         legend_xspan = 1.5
 
-    if plottype == 'multigroup' and float_contrast is True:
-        raw_xspan = 5
-        ws = 0.5
-    else:
-        if color_col is None:
-            raw_xspan = swarm_xspan + 1.5
-        else:
-            raw_xspan = swarm_xspan
+    # These values were obtained through aesthetic optimisation.
+    group_height = 8
+    float_aspect = 0.8
 
-    xinches = raw_xspan * ncols + legend_xspan
-    fsize = (xinches, 7)
+    if float_contrast is True:
+        height_inches = group_height * float_aspect
+        width_inches = 4 * ncols + legend_xspan
+    else:
+        height_inches = group_height
+        width_inches = 1.5 * ngroups + legend_xspan
+
+    fsize = (width_inches, height_inches)
 
     # Set figure size.
     if fig_size is None:
         fig_size = fsize
 
-
-    # Create subplots.
-    fig,axx=plt.subplots(ncols=ncols, figsize=fig_size,
-                         gridspec_kw={'width_ratios': widthratio,
-                                      'wspace' : ws}
-                        )
+    gridspecs = {'width_ratios': widthratio, 'wspace' : ws, 'hspace': 0}
+    subplots_args = {'ncols': ncols, 'figsize':fig_size, 'dpi':dpi,
+                     'gridspec_kw': gridspecs}
 
 
-    # If the contrast axes are NOT floating, create lists to store raw ylims
-    # and raw tick intervals, so that I can normalize their ylims later.
+    # Create the figure.
     if float_contrast is False:
+        fig, axx = plt.subplots(nrows=2, **subplots_args)
+        # If the contrast axes are NOT floating, create lists to store raw ylims
+        # and raw tick intervals, so that I can normalize their ylims later.
         contrast_ax_ylim_low = list()
         contrast_ax_ylim_high = list()
         contrast_ax_ylim_tickintervals = list()
 
+    else:
+        fig, axx = plt.subplots(nrows=1, **subplots_args)
 
     # CREATE COLOR PALETTE TO NORMALIZE PALETTE ACROSS AXES.
     if color_col is None:
@@ -547,6 +620,11 @@ def plot(data, idx,
 
     # FOR EACH TUPLE IN IDX, CREATE PLOT.
     for j, current_tuple in enumerate(idx):
+        # Check that current_tuple has no duplicate categories.
+        if len(np.unique(current_tuple)) != len(current_tuple):
+            err = '{} has duplicate categories.'.format(current_tuple)
+            raise ValueError(err)
+
         plotdat = data_in[data_in[x].isin(current_tuple)].copy()
         plotdat.loc[:,x] = pd.Categorical(plotdat[x],
                             categories=current_tuple,
@@ -556,33 +634,37 @@ def plot(data, idx,
         # Compute Ns per group.
         counts = plotdat.groupby(x)[y].count()
 
-        if ncols == 1:
-            ax_raw = axx
-        else:
-            ax_raw = axx[j]
-
         if float_contrast:
+            if ncols == 1:
+                ax_raw = axx
+            else:
+                ax_raw = axx[j]
             ax_contrast = ax_raw.twinx()
         else:
-            divider = make_axes_locatable(ax_raw)
-            ax_contrast = divider.append_axes("bottom", size="100%",
-                                            pad=0.5, sharex=ax_raw)
+            if ncols == 1:
+                ax_raw = axx[0]
+                ax_contrast = axx[1]
+            else:
+                ax_raw = axx[0, j] # the swarm axes are always on row 0.
+                ax_contrast = axx[1, j] # the contrast axes are always on row 1.
 
-        # PLOT RAW DATA.
         ax_raw.set_ylim(swarm_ylim)
 
+        # Plot the raw data.
         if (paired is True and show_pairs is True):
-            # first, sanity checks. Do we have 2 elements (no more, no less) here?
+            # Sanity checks. Do we have 2 elements (no more, no less) here?
             if len(current_tuple) != 2:
-                raise ValueError('Paired plotting is True, but {0} does not have 2 elements.'\
-                                 .format(str(current_tuple)) )
+                err1 = 'Paired plotting is True, '
+                err2 = 'but {0} does not have 2 elements.'.format(current_tuple)
+                raise ValueError(err1 + err2)
 
             # Are the groups equal in length??
             before = plotdat[plotdat[x] == current_tuple[0]][y].dropna().tolist()
             after = plotdat[plotdat[x] == current_tuple[1]][y].dropna().tolist()
             if len(before) != len(after):
-                raise ValueError('The sizes of {0} and {1} do not match.'\
-                                 .format(current_tuple[0], current_tuple[1]) )
+                err1 = 'The sizes of {} '.format(current_tuple[0])
+                err2 = 'and {} do not match.'.format(current_tuple[1])
+                raise ValueError(err1 + err2)
 
             if color_col is not None:
                 colors = plotdat[plotdat[x] == current_tuple[0]][color_col]
@@ -593,27 +675,28 @@ def plot(data, idx,
                                    str(current_tuple[1]):after,
                                    'colors':colors})
 
+            # Slopegraph for paired raw data points.
             for ii in linedf.index:
-                ax_raw.plot([0,1],  # x1, x2
+                ax_raw.plot([0, 1],  # x1, x2
                             [linedf.loc[ii,current_tuple[0]],
                              linedf.loc[ii,current_tuple[1]]] , # y1, y2
                             linestyle='solid', linewidth = 1,
-                            color = plotPal[linedf.loc[ii,'colors']],
-                            label = linedf.loc[ii,'colors'])
-            ax_raw.set_xticks([0,1])
+                            color = plotPal[linedf.loc[ii, 'colors']],
+                            label = linedf.loc[ii, 'colors'])
+            ax_raw.set_xticks([0, 1])
             ax_raw.set_xlim(-0.25, 1.5)
-            ax_raw.set_xticklabels([current_tuple[0],current_tuple[1]])
+            ax_raw.set_xticklabels([current_tuple[0], current_tuple[1]])
+            # swarm_ylim = ax_raw.get_ylim()
 
         elif (paired is True and show_pairs is False) or (paired is False):
-            # Swarmplot for raw data points.
-            sns.swarmplot(data=plotdat,
-                          x=x, y=y,
-                          ax=ax_raw,
-                          order=current_tuple,
-                          hue=color_col,
-                          palette=plotPal,
-                          zorder=3,
-                          **swarmplot_kwargs)
+            # # Swarmplot for raw data points.
+            # if swarm_ylim is not None:
+            #     ax_raw.set_ylim(swarm_ylim)
+            sns.swarmplot(data=plotdat, x=x, y=y, ax=ax_raw,
+                          order=current_tuple, hue=color_col,
+                          palette=plotPal, zorder=3, **swarmplot_kwargs)
+            # if swarm_ylim is None:
+            #     swarm_ylim = ax_raw.get_ylim()
 
             if group_summaries != 'None':
                 # Create list to gather xspans.
@@ -627,8 +710,8 @@ def plot(data, idx,
                         # we have got a None, so skip and move on.
                         pass
                 gapped_lines(plotdat, x=x, y=y,
-                             # pseudo-hardcorded offset...
-                             offset=np.max(xspans)+0.09,
+                             # Hardcoded offset...
+                             offset=np.max(xspans) + 0.1,
                              type=group_summaries,
                              ax=ax_raw, **group_summary_kwargs)
 
@@ -637,18 +720,17 @@ def plot(data, idx,
 
         # Set new tick labels. The tick labels belong to the SWARM axes
         # for both floating and non-floating plots.
-        # This is because `sharex` was invoked.
-        newticklabs = list()
+        xticklabels = list()
         for xticklab in ax_raw.xaxis.get_ticklabels():
             t = xticklab.get_text()
-            N = str(counts.ix[t])
+            N = str(counts.loc[t])
             if show_group_count:
-                newticklabs.append(t+' n='+N)
+                xticklabels.append(t+' n='+N)
             else:
-                newticklabs.append(t)
-            ax_raw.set_xticklabels(newticklabs,
-                rotation=45,
-                horizontalalignment='right')
+                xticklabels.append(t)
+            if float_contrast is True:
+                ax_raw.set_xticklabels(xticklabels, rotation=45,
+                                       horizontalalignment='right')
 
 
         # Despine appropriately.
@@ -712,10 +794,10 @@ def plot(data, idx,
                                        **violinplot_kwargs)
             halfviolin(v)
             ax_contrast.plot([pos+1], boots.summary, marker='o', color='k',
-                            markersize=swarmplot_kwargs['size'] * 1.5)
+                             markersize=difference_dotsize)
             ax_contrast.plot([pos+1,pos+1],
                              [boots.bca_ci_low, boots.bca_ci_high],
-                             'k-', linewidth=group_summary_kwargs['lw'])
+                             'k-', linewidth=ci_linewidth)
             if float_contrast is False:
                 contrast_ax_ylim_low.append( ax_contrast.get_ylim()[0] )
                 contrast_ax_ylim_high.append( ax_contrast.get_ylim()[1] )
@@ -724,8 +806,13 @@ def plot(data, idx,
 
 
 
-        # NORMALISE Y LIMS AND DESPINE FLOATING CONTRAST AXES.
-        if float_contrast:
+        if float_contrast is False:
+            ax_contrast.set_xlim(ax_raw.get_xlim())
+            ax_contrast.set_xticks(ax_raw.get_xticks())
+            ax_contrast.set_xticklabels(xticklabels, rotation=45,
+                                        horizontalalignment='right')
+
+        else: # float_contrast is True
             # Check that the effect size is within the swarm ylims.
             min_check = swarm_ylim[0] - ref.mean()
             max_check = swarm_ylim[1] - ref.mean()
@@ -916,9 +1003,15 @@ def plot(data, idx,
         fig.axes[0].set_ylabel(swarm_label)
 
 
-    # Lengthen the axes ticks so they look better.
+    axes_spine_linewidth = 1.25
+    # After some visual investigation, this seemed to be an ideal thickness.
     for ax in fig.axes:
-        ax.tick_params(length=tick_length, pad=tick_pad)
+        # Lengthen the axes ticks so they look better.
+        ax.tick_params(length=tick_length, pad=tick_pad,
+                       width=axes_spine_linewidth)
+        # Set a thinner axes spine for better aesthetics.
+        for axis in ['top','bottom','left','right']:
+            ax.spines[axis].set_linewidth(axes_spine_linewidth)
 
 
     # Return the figure and the results DataFrame.