+++ dabest/_plot_tools.py
@@ -10,28 +10,30 @@ import matplotlib.pyplot as plt
 import seaborn as sns
 import numpy as np
 
-from .misc_tools import merge_two_dicts
+from ._misc_tools import merge_two_dicts
 
 
 
-def halfviolin(v, half = 'right', color = 'k'):
+def halfviolin(v, half='right', color='k'):
+    import numpy as np
+
     for b in v['bodies']:
-            mVertical = np.mean(b.get_paths()[0].vertices[:, 0])
-            mHorizontal = np.mean(b.get_paths()[0].vertices[:, 1])
-            vertices = b.get_paths()[0].vertices
-            if half is 'left':
-                b.get_paths()[0].vertices[:, 0] = np.clip(vertices[:, 0],
-                                                    -np.inf, mVertical)
-            if half is 'right':
-                b.get_paths()[0].vertices[:, 0] = np.clip(vertices[:, 0],
-                                                    mVertical, np.inf)
-            if half is 'bottom':
-                b.get_paths()[0].vertices[:, 1] = np.clip(vertices[:, 1],
-                                                    -np.inf, mHorizontal)
-            if half is 'top':
-                b.get_paths()[0].vertices[:, 1] = np.clip(vertices[:, 1],
-                                                    mHorizontal, np.inf)
-            b.set_color(color)
+        V = b.get_paths()[0].vertices
+
+        mean_vertical = np.mean(V[:, 0])
+        mean_horizontal = np.mean(V[:, 1])
+
+        if half is 'left':
+            V[:, 0] = np.clip(V[:, 0], -np.inf, mean_vertical)
+        if half is 'right':
+            V[:, 0] = np.clip(V[:, 0], mean_vertical, np.inf)
+        if half is 'bottom':
+            V[:, 1] = np.clip(V[:, 1], -np.inf, mean_horizontal)
+        if half is 'top':
+            V[:, 1] = np.clip(V[:, 1], mean_horizontal, np.inf)
+
+        b.set_color(color)
+        b.set_linewidth(0)
 
 
 
@@ -109,6 +111,7 @@ def gapped_lines(data, x, y,
 
     if ax is None:
         ax = plt.gca()
+    ax_ylims = ax.get_ylim()
 
     keys = kwargs.keys()
     if 'zorder' not in keys:
@@ -140,6 +143,11 @@ def gapped_lines(data, x, y,
         lows = lower_quartiles
         highs = upper_quartiles
 
+    if (lows < ax_ylims[0]).any() or (highs > ax_ylims[1]).any():
+        kwargs['clip_on'] = True
+    else:
+        kwargs['clip_on'] = False
+
     original_zorder = kwargs['zorder']
     span_color = kwargs['color']
     span_lw = kwargs['lw']
@@ -156,7 +164,7 @@ def gapped_lines(data, x, y,
         # add horzontal central measure line.
         kwargs['zorder'] = 6
         kwargs['color'] = 'white'
-        kwargs['lw'] = span_lw + 0.5
+        kwargs['lw'] = span_lw * 1.5
         mean_line = mlines.Line2D([xpos+offset-0.01,
                                     xpos+offset+0.01],
                                     [cm, cm],