+++ docs/source/release-notes.rst
@@ -5,6 +5,20 @@ Release Notes
 =============
 
 
+v0.1.6
+======
+Several keywords have been added to allow more fine-grained control over a selection of plot elements.
+
+* `swarm_dotsize`
+* `difference_dotsize`
+* `ci_linewidth`
+* `summary_linewidth`
+
+The new keyword `context` allows you to set the plotting context as defined by seaborn's `plotting_context() <https://seaborn.pydata.org/generated/seaborn.plotting_context.html>`_ .
+
+Now, if `paired=True`, you will need to supply an `id_col`, which is a column in the DataFrame which specifies which sample the datapoint belongs to. See :ref:`producing-slopgraphs` for more details.
+
+
 v0.1.5
 ======
 Fix bug that wasn't updating the seaborn version upon setup and install.